Workaround package visibility gap in getArchivedPackage()
This is just one of the many exploits that AOSP suffers from. Since
this issue is fixed by me in AOSP, we might as well address it here
too. Please note: Only way to "reliably" achieve package visibility
is to install it in a separate user. HMA is just a workaround.